REFACTOR: configuration package, tests
