+++ src/main/java/es/in2/issuer/configuration/adapter/azure/config/AzureAppConfigurationConfig.java
@@ -1,4 +1,4 @@
-package es.in2.issuer.configuration.azure.config;
+package es.in2.issuer.configuration.adapter.azure.config;
 
 import com.azure.core.credential.TokenCredential;
 import com.azure.data.appconfiguration.ConfigurationClient;
@@ -6,7 +6,7 @@ import com.azure.data.appconfiguration.ConfigurationClientBuilder;
 import com.azure.identity.DefaultAzureCredentialBuilder;
 import es.in2.issuer.configuration.model.ConfigProviderName;
 import es.in2.issuer.configuration.util.ConfigSourceName;
-import es.in2.issuer.configuration.azure.config.properties.AzureProperties;
+import es.in2.issuer.configuration.adapter.azure.config.properties.AzureProperties;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Bean;
