+++ src/main/java/es/in2/issuer/configuration/adapter/azure/config/properties/AzurePropertiesLabel.java
@@ -1,4 +1,4 @@
-package es.in2.issuer.configuration.azure.config.properties;
+package es.in2.issuer.configuration.adapter.azure.config.properties;
 
 import es.in2.issuer.configuration.model.ConfigProviderName;
 import es.in2.issuer.configuration.util.ConfigSourceName;
