+++ src/test/java/es/in2/issuer/api/service/impl/RemoteSignatureServiceImplTest.java
@@ -1,14 +1,12 @@
-/*
 package es.in2.issuer.api.service.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import es.in2.issuer.api.config.azure.AppConfigurationKeys;
+import es.in2.issuer.api.config.AppConfiguration;
 import es.in2.issuer.api.model.dto.SignatureConfiguration;
 import es.in2.issuer.api.model.dto.SignatureRequest;
 import es.in2.issuer.api.model.dto.SignedData;
 import es.in2.issuer.api.model.enums.SignatureType;
-import es.in2.issuer.api.service.AppConfigService;
 import es.in2.issuer.api.util.HttpUtils;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -34,7 +32,7 @@ import static org.mockito.Mockito.*;
 class RemoteSignatureServiceImplTest {
 
     @Mock
-    private AppConfigService appConfigService;
+    private AppConfiguration appConfiguration;
 
     @Mock
     private ObjectMapper objectMapper;
@@ -47,14 +45,16 @@ class RemoteSignatureServiceImplTest {
 
     @Test
     void testInitializeRemoteSignatureBaseUrl() throws NoSuchMethodException, IllegalAccessException, InvocationTargetException {
-        lenient().when(appConfigService.getConfiguration(any())).thenReturn(Mono.just("dummyValue"));
+        //lenient().when(appConfigService.getConfiguration(any())).thenReturn(Mono.just("dummyValue"));
+        lenient().when(appConfiguration.getRemoteSignatureDomain()).thenReturn(String.valueOf(Mono.just("dummyValue")));
 
         Method privateMethod = RemoteSignatureServiceImpl.class.getDeclaredMethod("initializeRemoteSignatureBaseUrl");
         privateMethod.setAccessible(true);
 
         privateMethod.invoke(remoteSignatureService);
 
-        verify(appConfigService, times(1)).getConfiguration(AppConfigurationKeys.CROSS_REMOTE_SIGNATURE_BASE_URL_KEY);
+        //verify(appConfigService, times(1)).getConfiguration(AppConfigurationKeys.ISSUER_AUTHENTIC_SOURCES_BASE_URL_KEY);
+        verify(appConfiguration, times(1)).getRemoteSignatureDomain();
     }
 
     @Test
@@ -63,11 +63,11 @@ class RemoteSignatureServiceImplTest {
         Method privateMethod = RemoteSignatureServiceImpl.class.getDeclaredMethod("initializeRemoteSignatureBaseUrl");
         privateMethod.setAccessible(true);
 
-        lenient().when(appConfigService.getConfiguration(AppConfigurationKeys.CROSS_REMOTE_SIGNATURE_BASE_URL_KEY)).thenReturn(Mono.error(new RuntimeException("Simulated error")));
+        when(appConfiguration.getRemoteSignatureDomain()).thenAnswer(invocation -> Mono.error(new RuntimeException("Simulated error")));
 
         assertThrows(InvocationTargetException.class, () -> privateMethod.invoke(remoteSignatureService));
 
-        verify(appConfigService, times(1)).getConfiguration(AppConfigurationKeys.CROSS_REMOTE_SIGNATURE_BASE_URL_KEY);
+        verify(appConfiguration, times(1)).getRemoteSignatureDomain();
     }
 
     @Test
@@ -153,4 +153,3 @@ class RemoteSignatureServiceImplTest {
     }
 
 }
-*/
